Simplify methods and classes with high CRAP score. Regenerate autoload_classmap.php
